+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.6.0
 
 * Add support for fetching data from NSIDC0802 v2 from disk.
+* Update dependencies to support `earthaccess ~=0.14.0`.
+* Add functions to fetch TBs from NSIDC-0001, NSIDC-0080 via `earthaccess`.
 
 ## 0.5.0